Added expressions of form 'not (a or b)'
It is not adding expressions of the form (a or b) since that does not
really add anything.

Not only does it expand the tests the game gives you. This also means
the program recognizes that the shortest form of:
'not a and not b' -> 'not (a or b)'

Before it thought that that expression could not be made shorter.